backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==5.0.7
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
setuptools==70.3.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
